requirements.txt
bleach==6.2.0
feedparser==6.0.11
Jinja2==3.1.6
pyaml==25.5.0
python-dateutil==2.9.0.post0
requests==2.32.3